+++ src/Impl/Impl.jl
@@ -5,7 +5,7 @@ module Impl
 import ..SingleCellProjections as SCP
 
 import .SCP.SCPCore
-using .SCPCore: DataMatrix, Blocks
+using .SCPCore: DataMatrix, Blocks, check_kwargs
 
 import SCTransform
 import SingleCell10x