+++ src/projectables.jl
@@ -1,5 +1,5 @@
 """
-    SCP.project(onto, old => new, ...; kwargs...) -> Job
+    SCP.project(onto, old => new, ...) -> Job
 
 Projects a dataset onto another, while replacing old=>new. Multiple replacement pairs can be specified.
 (TODO: Describe projection properly.)
@@ -11,6 +11,6 @@ Given a force layout Job `fl`, we here project `proj_raw_counts` onto that force
 julia> SCP.project(fl, raw_counts=>proj_raw_counts)
 ```
 """
-function project(onto, args...; kwargs...)
-	Impl.create_project_job(onto, args...; kwargs...)
+function project(onto, args...)
+	Impl.create_project_job(onto, args...)
 end