+++ src/pseudobulk.jl
@@ -9,6 +9,7 @@ specified covariates. Returns a `DataMatrix` where each column is a pseudobulk s
 See also [`population_matrix`](@ref).
 """
 function pseudobulk(data, obs_covariate1, obs_covariates...; kwargs...)
+	check_kwargs(kwargs, :new_var_covariates, :delim, :var_id_colname, :obs_id_colname)
 	create_job(Preprocess(Impl.pseudobulk), data, obs_covariate1, obs_covariates...; kwargs...)
 end
 
@@ -26,5 +27,6 @@ covariates define the columns (samples/groups) and `new_var_covariates` define t
 See also [`pseudobulk`](@ref).
 """
 function population_matrix(obs, obs_covariate1, obs_covariates...; new_var_covariates, kwargs...)
+	check_kwargs(kwargs, :delim, :var_id_colname, :obs_id_colname)
 	create_job(Preprocess(Impl.population_matrix), obs, obs_covariate1, obs_covariates...; new_var_covariates, kwargs...)
 end